Add simple tests
