+++ src/test/java/seedu/algobase/logic/commands/AddCommandTest.java
@@ -4,33 +4,25 @@ import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.algobase.commons.core.GuiSettings;
 import seedu.algobase.model.AlgoBase;
-import seedu.algobase.model.Model;
 import seedu.algobase.model.ReadOnlyAlgoBase;
-import seedu.algobase.model.ReadOnlyUserPrefs;
-import seedu.algobase.model.plan.Plan;
 import seedu.algobase.model.problem.Problem;
 import seedu.algobase.testutil.ProblemBuilder;
 
 class AddCommandTest {
     @Test
-    public void constructor_nullPerson_throwsNullPointerException() {
+    public void constructor_nullProblem_throwsNullPointerException() {
         assertThrows(NullPointerException.class, () -> new AddCommand(null));
     }
 
     @Test
-    public void execute_personAcceptedByModel_addSuccessful() throws Exception {
-        ModelStubAcceptingPersonAdded modelStub = new ModelStubAcceptingPersonAdded();
+    public void execute_problemAcceptedByModel_addSuccessful() throws Exception {
+        ModelStubAcceptingProblemAdded modelStub = new ModelStubAcceptingProblemAdded();
         Problem validProblem = new ProblemBuilder().build();
 
         CommandResult commandResult = new AddCommand(validProblem).execute(modelStub);
@@ -40,94 +32,9 @@ class AddCommandTest {
     }
 
     /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAlgoBaseFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAlgoBaseFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addProblem(Problem problem) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAlgoBase(ReadOnlyAlgoBase newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAlgoBase getAlgoBase() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasProblem(Problem problem) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteProblem(Problem problem) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setProblem(Problem target, Problem editedProblem) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Plan> getFilteredPlanList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Problem> getFilteredProblemList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredProblemList(Predicate<Problem> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateSortedProblemList(Comparator<Problem> problemComparator) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that always accept the problem being added.
      */
-    private class ModelStubAcceptingPersonAdded extends ModelStub {
+    private class ModelStubAcceptingProblemAdded extends DefaultModelStub {
         final ArrayList<Problem> problemsAdded = new ArrayList<>();
 
         @Override
